+++ Realm/RLMSyncUser.mm
@@ -80,22 +80,22 @@ using namespace realm;
     return nil;
 }
 
-+ (void)logInWithCredential:(RLMSyncCredential *)credential
-              authServerURL:(NSURL *)authServerURL
-               onCompletion:(RLMUserCompletionBlock)completion {
-    [self logInWithCredential:credential
-                authServerURL:authServerURL
-                      timeout:30
-                 onCompletion:completion];
++ (void)logInWithCredentials:(RLMSyncCredentials *)credentials
+               authServerURL:(NSURL *)authServerURL
+                onCompletion:(RLMUserCompletionBlock)completion {
+    [self logInWithCredentials:credentials
+                 authServerURL:authServerURL
+                       timeout:30
+                  onCompletion:completion];
 }
 
-+ (void)logInWithCredential:(RLMSyncCredential *)credential
-              authServerURL:(NSURL *)authServerURL
-                    timeout:(NSTimeInterval)timeout
-               onCompletion:(RLMUserCompletionBlock)completion {
++ (void)logInWithCredentials:(RLMSyncCredentials *)credentials
+               authServerURL:(NSURL *)authServerURL
+                     timeout:(NSTimeInterval)timeout
+                onCompletion:(RLMUserCompletionBlock)completion {
     RLMSyncUser *user = [[RLMSyncUser alloc] initWithAuthServer:authServerURL];
     [RLMSyncUser _performLogInForUser:user
-                           credential:credential
+                          credentials:credentials
                         authServerURL:authServerURL
                               timeout:timeout
                       completionBlock:completion];
@@ -192,7 +192,7 @@ using namespace realm;
 }
 
 + (void)_performLogInForUser:(RLMSyncUser *)user
-                  credential:(RLMSyncCredential *)credential
+                 credentials:(RLMSyncCredentials *)credentials
                authServerURL:(NSURL *)authServerURL
                      timeout:(NSTimeInterval)timeout
              completionBlock:(RLMUserCompletionBlock)completion {
@@ -204,19 +204,19 @@ using namespace realm;
         });
     };
 
-    // Special credential login should be treated differently.
-    if (credential.provider == RLMIdentityProviderAccessToken) {
-        [self _performLoginForDirectAccessTokenCredential:credential user:user completionBlock:theBlock];
+    // Special credentials login should be treated differently.
+    if (credentials.provider == RLMIdentityProviderAccessToken) {
+        [self _performLoginForDirectAccessTokenCredentials:credentials user:user completionBlock:theBlock];
         return;
     }
 
     // Prepare login network request
     NSMutableDictionary *json = [@{
-                                   kRLMSyncProviderKey: credential.provider,
-                                   kRLMSyncDataKey: credential.token,
+                                   kRLMSyncProviderKey: credentials.provider,
+                                   kRLMSyncDataKey: credentials.token,
                                    kRLMSyncAppIDKey: [RLMSyncManager sharedManager].appID,
                                    } mutableCopy];
-    NSMutableDictionary *info = [(credential.userInfo ?: @{}) mutableCopy];
+    NSMutableDictionary *info = [(credentials.userInfo ?: @{}) mutableCopy];
 
     if ([info count] > 0) {
         // Munge user info into the JSON request.
@@ -261,12 +261,12 @@ using namespace realm;
                                  completion:handler];
 }
 
-+ (void)_performLoginForDirectAccessTokenCredential:(RLMSyncCredential *)credential
-                                               user:(RLMSyncUser *)user
-                                    completionBlock:(nonnull RLMUserCompletionBlock)completion {
-    user.directAccessToken = credential.token;
-    NSString *identity = credential.userInfo[kRLMSyncIdentityKey];
-    NSAssert(identity != nil, @"Improperly created direct access token credential.");
++ (void)_performLoginForDirectAccessTokenCredentials:(RLMSyncCredentials *)credentials
+                                                user:(RLMSyncUser *)user
+                                     completionBlock:(nonnull RLMUserCompletionBlock)completion {
+    user.directAccessToken = credentials.token;
+    NSString *identity = credentials.userInfo[kRLMSyncIdentityKey];
+    NSAssert(identity != nil, @"Improperly created direct access token credentials.");
     user.identity = identity;
     RLMSyncUser *existingUser = [[RLMSyncManager sharedManager] _registerUser:user];
     RLMSyncUser *actualUser = existingUser ?: user;